finish account test
